requirements.txt
streamlit
weaviate-client
llama-index
llama-index-vector-stores-weaviate
llama-index-llms-openai
openai
python-dotenv